backend/app/services/claude.py
import os
from typing import Dict, List, Optional
from anthropic import Anthropic
from dotenv import load_dotenv

load_dotenv()


class ClaudeClient:
    """
    Placeholder: Claude API client for email classification and analysis.

    This client will use Claude to classify emails into categories,
    extract action items, and analyze urgency.
    """

    def __init__(self):
        self.api_key = os.getenv("ANTHROPIC_API_KEY")
        self.client = Anthropic(api_key=self.api_key) if self.api_key else None

    async def classify_email(
        self,
        subject: str,
        body: str,
        from_address: str,
        categories: List[Dict]
    ) -> Dict:
        """
        Placeholder: Classify an email into one of the predefined categories.

        Args:
            subject: Email subject
            body: Email body text
            from_address: Sender email address
            categories: List of available category dictionaries

        Returns:
            Dictionary with category_id, confidence, and reasoning

        TODO: Implement Claude API call with structured output
        """
        # TODO: Build prompt with categories and email content
        # TODO: Call Claude API with tool use for structured classification
        # TODO: Return category_id, confidence score, and reasoning

        return {
            "category_id": None,
            "confidence": 0.0,
            "reasoning": "Classification not yet implemented"
        }

    async def extract_action_items(self, subject: str, body: str) -> List[Dict]:
        """
        Placeholder: Extract action items from email content.

        Args:
            subject: Email subject
            body: Email body text

        Returns:
            List of action item dictionaries with description and due_date

        TODO: Implement Claude API call to extract action items
        """
        # TODO: Call Claude API to identify tasks and deadlines
        return []

    async def analyze_urgency(
        self,
        subject: str,
        body: str,
        from_address: str,
        category: str
    ) -> Dict:
        """
        Placeholder: Analyze email urgency using Claude.

        Args:
            subject: Email subject
            body: Email body text
            from_address: Sender email address
            category: Classified category

        Returns:
            Dictionary with urgency_score and factors

        TODO: Implement Claude API call for urgency analysis
        """
        # TODO: Build prompt analyzing urgency signals
        # TODO: Call Claude API
        # TODO: Return urgency score (0.0 to 1.0) and contributing factors

        return {
            "urgency_score": 0.0,
            "factors": []
        }

    async def batch_classify(self, emails: List[Dict]) -> List[Dict]:
        """
        Placeholder: Classify multiple emails in a single request.

        Args:
            emails: List of email dictionaries

        Returns:
            List of classification results

        TODO: Implement batch classification for efficiency
        """
        # TODO: Use Claude's batch API or prompt caching for efficiency
        return []
