backend/requirements.txt
fastapi
uvicorn
sqlalchemy
anthropic
msal
httpx
python-dotenv
